+++ xtask/src/bootstrap.rs
@@ -1202,7 +1202,7 @@ mod tests {
         );
         assert_eq!(
             manifest.bootstrap_compiler.tree_digest,
-            "sha256:0a84ac381ace43b4d7be650f1082aedd1bfc92c10002b50d2d6b223fef6b7ebc"
+            "sha256:085e030264fd56e3232ce0fcef480dc2b3f0fe79bb41d8bb7b7823d2a026ded8"
         );
         assert_eq!(manifest.bootstrap_compiler.lexer_snapshot_schema_version, 1);
         assert_eq!(manifest.bootstrap_compiler.differential_case_count, 11);